Add type (#48)
* 个人资产过滤条件bug

* add expire history

* add is_shelf sql

* use goto

* bug

* bug

* bug

* bug

* add_type:fix tx dead lock

* market not login

+++ services/web/portal_service.go
@@ -752,7 +752,6 @@ func (p portalServiceImp) UserHistory(info request.UserHistory) (out response.Us
 		out.PrePageCount = info.PageCount
 		out.Total = count
 		return out, 0, nil
-
 	case common.ListenHistory:
 	default:
 		slog.Slog.ErrorF(info.Ctx, "PlatformServiceImp UserHistory error:history type not exists")
@@ -816,7 +815,7 @@ func (p portalServiceImp) AssetDetail(info request.AssetDetail) (out response.As
 		Subcategory:     subCategoryString,
 		Status:          assetsOrders.Status,
 		Price:           assetsOrders.Price,
-		OrderId:         int64(assetsOrders.OrderID),
+		OrderId:         assetsOrders.OrderID,
 		ExpireTime:      assetsOrders.ExpireTime,
 		Signature:       assetsOrders.Signature,
 		StartTime:       assetsOrders.StartTime,
